Rename SAML2\Request to SAML2\XML\samlp\AbstractRequest
