+++ tests/SAML2/Response/SignatureValidationTest.php
@@ -4,15 +4,17 @@ declare(strict_types=1);
 
 namespace SAML2\Response;
 
-use SAML2\CertificatesMock;
+use DOMDocument;
+use Mockery;
 use SAML2\Assertion;
+use SAML2\CertificatesMock;
 use SAML2\Configuration\Destination;
 use SAML2\Configuration\IdentityProvider;
 use SAML2\Configuration\ServiceProvider;
-use SAML2\Response;
+use SAML2\Response\Exception\UnsignedResponseException;
 use SAML2\Utilities\ArrayCollection;
 use SAML2\Utilities\Certificate;
-use SAML2\Response\Exception\UnsignedResponseException;
+use SAML2\XML\samlp\Response;
 
 /**
  * Test that ensures that either the response or the assertion(s) or both must be signed.
@@ -52,8 +54,8 @@ class SignatureValidationTest extends \Mockery\Adapter\Phpunit\MockeryTestCase
      */
     public function setUp(): void
     {
-        $this->assertionProcessorBuilder = \Mockery::mock('alias:SAML2\Assertion\ProcessorBuilder');
-        $this->assertionProcessor = \Mockery::mock(Assertion\Processor::class);
+        $this->assertionProcessorBuilder = Mockery::mock('alias:SAML2\Assertion\ProcessorBuilder');
+        $this->assertionProcessor = Mockery::mock(Assertion\Processor::class);
         $this->assertionProcessorBuilder
             ->shouldReceive('build')
             ->once()
@@ -150,7 +152,7 @@ class SignatureValidationTest extends \Mockery\Adapter\Phpunit\MockeryTestCase
     {
         $this->expectException(UnsignedResponseException::class);
 
-        $assertion = \Mockery::mock('SAML2\Assertion');
+        $assertion = Mockery::mock('SAML2\Assertion');
 
         // The processAssertions is called to decrypt possible encrypted assertions,
         // after which it should fail with an exception due to having no signature
@@ -172,29 +174,29 @@ class SignatureValidationTest extends \Mockery\Adapter\Phpunit\MockeryTestCase
 
 
     /**
-     * @return \SAML2\Response
+     * @return \SAML2\XML\samlp\Response
      */
     private function getSignedResponseWithUnsignedAssertion(): Response
     {
-        $doc = new \DOMDocument();
+        $doc = new DOMDocument();
         $doc->load(__DIR__ . '/response.xml');
-        $response = new Response($doc->firstChild);
+        $response = Response::fromXML($doc->firstChild);
         $response->setSigningKey(CertificatesMock::getPrivateKey());
         $response->setCertificates([CertificatesMock::PUBLIC_KEY_PEM]);
 
         // convert to signed response
-        return new Response($response->toSignedXML());
+        return Response::fromXML($response->toSignedXML());
     }
 
 
     /**
-     * @return \SAML2\Response
+     * @return \SAML2\XML\samlp\Response
      */
     private function getUnsignedResponseWithSignedAssertion(): Response
     {
-        $doc = new \DOMDocument();
+        $doc = new DOMDocument();
         $doc->load(__DIR__ . '/response.xml');
-        $response = new Response($doc->firstChild);
+        $response = Response::fromXML($doc->firstChild);
 
         $assertions = $response->getAssertions();
         $assertion = $assertions[0];
@@ -209,13 +211,13 @@ class SignatureValidationTest extends \Mockery\Adapter\Phpunit\MockeryTestCase
 
 
     /**
-     * @return \SAML2\Response
+     * @return \SAML2\XML\samlp\Response
      */
     private function getSignedResponseWithSignedAssertion(): Response
     {
-        $doc = new \DOMDocument();
+        $doc = new DOMDocument();
         $doc->load(__DIR__ . '/response.xml');
-        $response = new Response($doc->firstChild);
+        $response = Response::fromXML($doc->firstChild);
         $response->setSigningKey(CertificatesMock::getPrivateKey());
         $response->setCertificates([CertificatesMock::PUBLIC_KEY_PEM]);
 
@@ -224,17 +226,17 @@ class SignatureValidationTest extends \Mockery\Adapter\Phpunit\MockeryTestCase
         $assertion->setSigningKey(CertificatesMock::getPrivateKey());
         $assertion->setCertificates([CertificatesMock::PUBLIC_KEY_PEM]);
 
-        return new Response($response->toSignedXML());
+        return Response::fromXML($response->toSignedXML());
     }
 
 
     /**
-     * @return \SAML2\Response
+     * @return \SAML2\XML\samlp\Response
      */
     private function getUnsignedResponseWithUnsignedAssertion(): Response
     {
-        $doc = new \DOMDocument();
+        $doc = new DOMDocument();
         $doc->load(__DIR__ . '/response.xml');
-        return new Response($doc->firstChild);
+        return Response::fromXML($doc->firstChild);
     }
 }
